requirements.txt
aerich==0.8.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==24.3.0
babel==2.16.0
bitarray==3.0.0
cached-property==2.0.1
cbor2==5.6.5
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
coverage==7.6.10
cytoolz==1.0.1
dictdiffer==0.9.0
eth-account==0.13.4
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-stdlib==0.2.8
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.6
ghp-import==2.1.0
h11==0.14.0
hexbytes==1.3.0
hypothesis==6.124.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
iso8601==2.1.0
Jinja2==3.1.5
lru-dict==1.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.41
mkdocs-material-extensions==1.3.1
packaging==23.2
paginate==0.5.7
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
py-ecc==7.0.1
py-evm==0.10.1b2
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pymdown-extensions==10.14.1
pypika-tortoise==0.5.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
titanoboa @ git+https://github.com/vyperlang/titanoboa.git@a8d5a6cd61632dbcb8f7c472974efa72decbe1b0
toolz==1.0.0
tortoise-orm==0.24.0
trie==3.0.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
vvm==0.3.2
vyper==0.4.0
watchdog==6.0.0
wheel==0.45.1
zipp==3.21.0
